Changes for 2.2.9 build. (#58)
* Add the Cavenic Shroom and improve the Cavenic Skeleton AI. (#51)

* add caver shroom(not generate yet) and cavenic skeleton ai changed

* changed follow range

* undo cavenic skeleton ai and add cavershroom function

* add generation

* Rename shroom

* Refactoring

* Changes for gradle

* Add coupon to Mirage Book.

* Add Durang Hog (#54)

* add caver shroom(not generate yet) and cavenic skeleton ai changed

* changed follow range

* undo cavenic skeleton ai and add cavershroom function

* add generation

* add durang hog

* change durang hog texture

* some fixed

* some fixed

* code fixed

* delete old file

* Update en_us.lang

* Delete caver_shroom.json


* Refactoring

* Add shroom gen to Huge Cavern

* Update version to 2.2.7

* Fix shroom gen rate

* Add crazy spawn chance

* Fix cavenia spawn

* Add custom biome layer

* Fix chunk gen

* Fix tower dungeon gen lag

* Fix overlapped spawn logic

* Bump version

* Fix biome gen weight

* Refactoring

* Refactoring network

* Fix spawn check pos

* Fix not spawn aqua squids

* Fix dark library gen

* Add actual mining

* Bump version

+++ src/main/java/cavern/util/BlockMeta.java
@@ -240,7 +240,7 @@ public class BlockMeta implements Comparable<BlockMeta>
 							{
 								try
 								{
-									return ((Integer)method.invoke(obj, new Object[0])).intValue();
+									return (Integer)method.invoke(obj, new Object[0]);
 								}
 								catch (Exception e) {}
 							}
@@ -259,7 +259,7 @@ public class BlockMeta implements Comparable<BlockMeta>
 		public String load(Pair<Block, Integer> key) throws Exception
 		{
 			Block block = key.getLeft();
-			int meta = key.getRight().intValue();
+			int meta = key.getRight();
 
 			if (block == null)
 			{
